refact: Leaflet to maps route as container
